Move observability options into a separate file
